Move tables code, more progress
