+++ requirements.txt
@@ -20,7 +20,7 @@ PyYAML==6.0.1
 uritemplate==4.1.1
 crispy-bootstrap4==2023.1
 django-plotly-dash==2.2.2
-dash-ag-grid
+dash-ag-grid==2.4.0
 
 ## Legacy dependency versions
 # asgiref==3.3.4
